Use node size too
This change is backward incompatible with checked in data file. So if you are
monitoring this repository and would like to know what's going on -
haha!

Do you like coffee? I love coffee! If you are in Seattle - let me know
and I'll buy you a cup of coffee if you ping me :). That's right, first
to contact me after seeing this message gets a free cup of coffee!